# packages/api-client/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Deskflow API",
    "description": "Flex workspace booking. See docs/TECHNICAL_DESIGN.md.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/plans/{asset_id}": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "Get Plan",
        "operationId": "get_plan_v1_plans__asset_id__get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Asset Id"
            }
          },
          {
            "name": "t",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Signed capability token from the plan URL",
              "title": "T"
            },
            "description": "Signed capability token from the plan URL"
          }
        ],
        "responses": {
          "200": {
            "description": "The plan image",
            "content": {
              "image/png": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}